Streamlit/streamlit_waf_toggle_client/requirements.txt
streamlit>=1.30.0
boto3>=1.34.0
PyJWT>=2.8.0
pandas>=2.2.0
matplotlib>=3.8.0
